# Conflicts:
#	audace-backend/src/main/java/com/equipe4/audace/AudaceApplication.java
#	audace-backend/src/main/java/com/equipe4/audace/controller/EmployerController.java
#	audace-backend/src/main/java/com/equipe4/audace/controller/StudentController.java
#	audace-backend/src/main/java/com/equipe4/audace/dto/application/ApplicationDTO.java
#	audace-backend/src/main/java/com/equipe4/audace/model/application/Application.java
#	audace-backend/src/main/java/com/equipe4/audace/repository/application/ApplicationRepository.java
#	audace-backend/src/main/java/com/equipe4/audace/repository/offer/OfferRepository.java
#	audace-backend/src/main/java/com/equipe4/audace/service/EmployerService.java
#	audace-backend/src/main/java/com/equipe4/audace/service/StudentService.java
#	audace-backend/src/test/java/com/equipe4/audace/controller/EmployerControllerTest.java
#	audace-backend/src/test/java/com/equipe4/audace/controller/StudentControllerTest.java
#	audace-backend/src/test/java/com/equipe4/audace/service/EmployerServiceTest.java
#	audace-backend/src/test/java/com/equipe4/audace/service/StudentServiceTest.java
#	audace-frontend/src/App.tsx
#	audace-frontend/src/components/OffersList/OfferRow/OfferButtons/ManagerButtons/index.tsx
#	audace-frontend/src/components/OffersList/OfferRow/OfferButtons/StudentButtons/index.tsx
#	audace-frontend/src/components/OffersList/OfferRow/OfferModal/index.tsx
#	audace-frontend/src/services/applicationService.ts
#	audace-frontend/src/services/cvService.ts
#	audace-frontend/src/services/offerService.ts
#	audace-frontend/src/services/studentApplicationService.ts
#	audace-frontend/src/views/Employer/EmployerApplicationView/index.tsx
#	audace-frontend/src/views/Employer/EmployerOfferView/index.tsx
#	audace-frontend/src/views/Employer/EmployerView/index.tsx
#	audace-frontend/src/views/Student/StudentView/index.tsx

+++ audace-backend/src/main/java/com/equipe4/audace/service/EmployerService.java
@@ -6,10 +6,14 @@ import com.equipe4.audace.dto.offer.OfferDTO;
 import com.equipe4.audace.model.Employer;
 import com.equipe4.audace.model.application.Application;
 import com.equipe4.audace.model.offer.Offer;
+import com.equipe4.audace.model.session.OfferSession;
+import com.equipe4.audace.model.session.Session;
 import com.equipe4.audace.repository.EmployerRepository;
 import com.equipe4.audace.repository.application.ApplicationRepository;
 import com.equipe4.audace.repository.offer.OfferRepository;
+import com.equipe4.audace.repository.session.OfferSessionRepository;
 import com.equipe4.audace.repository.security.SaltRepository;
+import com.equipe4.audace.utils.SessionManipulator;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
@@ -20,11 +24,22 @@ public class EmployerService extends GenericUserService<Employer> {
     private final EmployerRepository employerRepository;
     private final OfferRepository offerRepository;
     private final ApplicationRepository applicationRepository;
+    private final OfferSessionRepository offerSessionRepository;
+    private final SessionManipulator sessionManipulator;
 
-    public EmployerService(SaltRepository saltRepository, EmployerRepository employerRepository, OfferRepository offerRepository, ApplicationRepository applicationRepository) {
+    public EmployerService(
+            SaltRepository saltRepository,
+            EmployerRepository employerRepository,
+            OfferRepository offerRepository,
+            OfferSessionRepository offerSessionRepository,
+            SessionManipulator sessionManipulator,
+            ApplicationRepository applicationRepository
+    ) {
         super(saltRepository);
         this.employerRepository = employerRepository;
         this.offerRepository = offerRepository;
+        this.offerSessionRepository = offerSessionRepository;
+        this.sessionManipulator = sessionManipulator;
         this.applicationRepository = applicationRepository;
     }
 
@@ -49,28 +64,51 @@ public class EmployerService extends GenericUserService<Employer> {
         return employerRepository.findAll().stream().map(Employer::toDTO).toList();
     }
 
+    @Transactional
     public Optional<OfferDTO> createOffer(OfferDTO offerDTO){
         if(offerDTO == null) throw new IllegalArgumentException("Offer cannot be null");
 
+        Session session = sessionManipulator.getCurrentSession();
+
         Offer offer = offerDTO.fromDTO();
         offer.setOfferStatus(Offer.OfferStatus.PENDING);
+        offer = offerRepository.save(offerDTO.fromDTO());
 
+        offerSessionRepository.save(new OfferSession(null, offer, session));
         return Optional.of(offerRepository.save(offer).toDTO());
+        return Optional.of(offer.toDTO());
     }
 
-    public List<OfferDTO> findAllOffersByEmployerId(Long employerId){
-        Employer employer = employerRepository.findById(employerId).orElseThrow(() -> new NoSuchElementException("Employer not found"));
-        return offerRepository.findAllByEmployer(employer).stream().map(Offer::toDTO).toList();
+    @Transactional
+    public List<OfferDTO> findAllOffersByEmployerId(Long employerId, Long sessionId){
+        Employer employer = employerRepository.findById(employerId).orElseThrow();
+        List<Offer> offers = offerRepository.findAllByEmployer(employer);
+
+        return sessionManipulator.removeOffersNotInSession(offers, sessionId).stream().map(Offer::toDTO).toList();
     }
 
+    @Transactional
     public Optional<OfferDTO> updateOffer(OfferDTO offerDTO){
         Offer offer = offerRepository.findById(offerDTO.getId()).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        if (!sessionManipulator.isOfferInCurrentSession(offer)) {
+            throw new IllegalStateException("Offer is not in current session");
+        }
         return Optional.of(offerRepository.save(offer).toDTO());
     }
 
+    @Transactional
     public void deleteOffer(Long offerId){
         Offer offer = offerRepository.findById(offerId).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        if (!sessionManipulator.isOfferInCurrentSession(offer)) {
+            throw new IllegalStateException("Offer is not in current session");
+        }
         offerRepository.delete(offer);
+    }
+
+    //TODO : TESTS
+    public List<ApplicationDTO> findAllApplicationsByEmployerIdAndOfferId(Long employerId, Long offerId) {
+        Offer offer = offerRepository.findByEmployerIdAndId(employerId, offerId).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        return applicationRepository.findAllByOffer(offer).stream().map(Application::toDTO).toList();
     }
 
     public Map<Long, List<ApplicationDTO>> findAllApplicationsByEmployerId(Long employerId){
